add resource manager and update enum to enum class

+++ src/events/action.cpp
@@ -16,10 +16,10 @@ bool Action::test() const {
   bool res = false;
 
   if (event_.type == sf::Event::EventType::KeyPressed) {
-    if (type_ & Type::Pressed)
+    if (type_ & static_cast<int>(Type::Pressed))
       res = sf::Keyboard::isKeyPressed(event_.key.code);
   } else if (event_.type == sf::Event::EventType::MouseButtonPressed) {
-    if (type_ & Type::Pressed)
+    if (type_ & static_cast<int>(Type::Pressed))
       res = sf::Mouse::isButtonPressed(event_.mouseButton.button);
   }
   return res;
@@ -30,28 +30,28 @@ bool Action::operator==(const sf::Event& event) const {
 
   switch (event.type) {
     case sf::Event::EventType::KeyPressed:
-      if ((type_ & Type::Pressed) &&
+      if ((type_ & static_cast<int>(Type::Pressed)) &&
           (event_.type == sf::Event::EventType::KeyPressed))
         res = event.key.code == event_.key.code;
 
       break;
 
     case sf::Event::EventType::KeyReleased:
-      if ((type_ & Type::Released) &&
+      if ((type_ & static_cast<int>(Type::Released)) &&
           (event_.type == sf::Event::EventType::KeyPressed))
         res = event.key.code == event_.key.code;
 
       break;
 
     case sf::Event::EventType::MouseButtonPressed:
-      if ((type_ & Type::Pressed) &&
+      if ((type_ & static_cast<int>(Type::Pressed)) &&
           (event_.type == sf::Event::EventType::MouseButtonPressed))
         res = event.mouseButton.button == event_.mouseButton.button;
 
       break;
 
     case sf::Event::EventType::MouseButtonReleased:
-      if ((type_ & Type::Released) &&
+      if ((type_ & static_cast<int>(Type::Released)) &&
           (event_.type == sf::Event::EventType::MouseButtonPressed))
         res = event.mouseButton.button == event_.mouseButton.button;
